Fixed FindBugs warning

+++ sketches/src/main/java/com/yahoo/sketches/ArrayOfLongsSerDe.java
@@ -30,7 +30,7 @@ public class ArrayOfLongsSerDe extends ArrayOfItemsSerDe<Long> {
 
   @Override
   public Long[] deserializeFromMemory(final Memory mem, final int length) {
-    UnsafeUtil.checkBounds(0, length * Long.BYTES, mem.getCapacity());
+    UnsafeUtil.checkBounds(0, (long)length * Long.BYTES, mem.getCapacity());
     final Long[] array = new Long[length];
     long offsetBytes = 0;
     for (int i = 0; i < length; i++) {
